Complete unit testing (#7)
complete unit testing for usecases for whole project .

+++ app/Features/Categories/Application/Usecases/DestroyCategoryUsecase.php
@@ -31,6 +31,12 @@ final class DestroyCategoryUsecase implements DestroyCategoryContract
                 $presenter->onNotFound();
                 return;
             }
+            if ($record->isDefaultGroup) {
+                $presenter->onDefaultGroup();
+                return;
+            }
+            //---!!!!! move documents from the delete category to default caetgory
+            // before delete it !!!!! ------
             $this->categoryRepository->destroy($categoryId);
             $presenter->onSuccess();
         } catch (Exception $e) {
